% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visit_lines.R
\name{create_visit_lines}
\alias{create_visit_lines}
\title{Create two visit lines}
\usage{
create_visit_lines(chartcode, yname, period = numeric(0))
}
\arguments{
\item{chartcode}{The chart code, usually constructed by \code{chartbox::create_chartcode()}}

\item{yname}{Character indicating the measure}

\item{period}{A vector of length 2 with left and right ages, or \code{numeric(0)} (default) if there are
no ages}
}
\value{
An object of class \code{grob} with the visit lines
}
\description{
Create two visit lines
}
