% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matches.R
\name{find_matches}
\alias{find_matches}
\title{Find matches on all outcomes shown on the active chart}
\usage{
find_matches(
  individual,
  con,
  dnr,
  ynames,
  nmatch = 5L,
  period,
  user_model = 2L,
  exact_sex = FALSE,
  exact_ga = FALSE,
  break_ties = TRUE
)
}
\arguments{
\item{individual}{An S4 object of class \code{individual}
containing data of the individual}

\item{con}{A connection on which the donor data reside. The
default (\code{NULL}) reads from \code{donorloader} package.}

\item{dnr}{A string with the name of the donor data
(currently available are \code{smocc}, \code{terneuzen},
\code{lollypop} and \code{pops})}

\item{ynames}{A vector with the names of the response variables
for which matches are sought, e.g. \code{ynames = c("hdc",
"hgt")}.}

\item{nmatch}{Integer. Number of matches needed. When
\code{nmatch == 0L} no matches are sought.}

\item{period}{A vector of length 2 with left and right ages
(decimal age). If \code{length(period) == 0L}, then no curve
matching is done}

\item{user_model}{Model number (1-4), indicating type of model the
user wants. See details.}

\item{exact_sex}{A logical indicating whether sex should be
matched exactly}

\item{exact_ga}{A logical indicating whether gestational age
should be matched exactly}

\item{break_ties}{A logical indicating whether ties should broken
randomly. The default (\code{TRUE}) breaks ties randomly.}
}
\value{
List with components with names \code{ynames}. Each
  component is a vector of integers representing the id's of the
  matches cases from \code{donor}. Returns \code{integer(0)} if no
  matches are found.
}
\description{
Find matches on all outcomes shown on the active chart
}
\seealso{
\code{\link[curvematching]{calculate_matches}},
  \code{\link[curvematching]{extract_matches}}
}
