% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenband.R
\name{greenband}
\alias{greenband}
\title{Draw curves on green band}
\usage{
greenband(target, visible = NULL, design = c("A", "B", "C"), ratio = 3/4, ...)
}
\arguments{
\item{target}{A list with elements \code{"psn"} and \code{"xyz"}, e.g. as produced by
\code{bdsreader::read_bds()}}

\item{visible}{A named list specifying legend entries, e.g. as
\code{visible = list(hgt = TRUE, hdc = "legendonly")}. The three choices are:
\code{TRUE} (curve and legend shown), \code{FALSE} (curve and legend not shown)
and \code{"legendonly"} (curve not shown, legend dimmed). Unspecified entries are
set to \code{FALSE} and hence not shown. The default (\code{NULL}) sets curve and legend
visibility according to the \code{design} parameter.}

\item{design}{Either "A" (0-15m), "B" (0-4y) or "C" (0-21y)}

\item{ratio}{height/width ratio of the figure size. The default is 3/4.}

\item{\dots}{Passed down to plotting functions}
}
\value{
A \code{plotly} object
}
\description{
Draw curves on green band
}
\examples{
\dontrun{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
target <- bdsreader::read_bds(fn)
fig <- greenband(target)
fig
fig <- greenband(target, visible = list(hgt = "legendonly", dsc = TRUE, wgt = FALSE))
fig
}
}
