% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlines.R
\name{gridlinesGrob}
\alias{gridlinesGrob}
\alias{grid.gridlines}
\title{Draw gridlines for axes object}
\usage{
gridlinesGrob(
  xy.axes,
  at = list(NULL, NULL),
  lwd = 1,
  col = 2,
  lty = 1,
  vp = NULL,
  name = "gridlines",
  ...
)

grid.gridlines(...)
}
\arguments{
\item{xy.axes}{The axes definition produced by \code{chartdesigner:::axesGrob()$xy.axes}}

\item{at}{A list of length 2 containing the locations where the grid lines
should be drawn. The first list component contains the values on the \code{x} axis
(for vertical grid lines). The second component the values on the \code{y} axis
(for horizontal grid lines).}

\item{lwd}{Scalar or list of length 2 containig line width for grid lines}

\item{col}{Scalar or list of length 2 containing color if the grid lines}

\item{lty}{Scaler or list of length 2 containing the \code{lty} lint type parameter}

\item{vp}{Viewport to draw to}

\item{name}{name of the \code{grob}}

\item{\dots}{Additional argument passed to \code{gpar()}}
}
\value{
A \code{gTree} that contains the gridlines grobs, under child names
\code{vertical} and \code{horizontal}. A \code{NULL} component does not draw grid lines.
}
\description{
Return a grob object that contains graphic object of four axes.
}
