% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sds.R
\name{init_sds}
\alias{init_sds}
\title{Create empty SDS plot}
\usage{
init_sds(
  design = c("A", "B", "C"),
  language = c("dutch", "english"),
  ynames = "",
  plot_library = c("plotly", "ggplot2"),
  ...
)
}
\arguments{
\item{design}{Either "A" (0-15m), "B" (0-4y) or "C" (0-21y)}

\item{language}{Either "dutch" or "english". Only "dutch" is currently
supported}

\item{ynames}{Character vector with \code{yname}. Needed only for ggplot2 to set the
legend with active trace lines.}

\item{plot_library}{Either "ggplot2" or "plotly"}

\item{\dots}{Argument passed down (\code{width = 16} and \code{height = 12}) to fixate the
aspect ratio.}
}
\description{
Create empty SDS plot
}
