% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xname.R
\name{make_xname}
\alias{make_xname}
\title{Creates the model specifications for xname}
\usage{
make_xname(yname, xnames, user_model, current_age)
}
\arguments{
\item{yname}{String indicating the dependent variable}

\item{xnames}{Character vector of available names to choose from.
It is assumed all variables named in \code{xnames} have complete
data. Any brokenstick variable should have names like
\code{paste(yname, d, sep = "_")}, where \code{d} is a vector
contuining decimal ages.}

\item{user_model}{Model number (1-4), indicating type of model the user wants.
See details.}

\item{current_age}{Current age (in decimal years)}
}
\value{
A character vector with the \code{xname} specification
}
\description{
Creates the model specifications for xname
}
\examples{
xnames <- c("sex", "ga", "bw", "hgt_0", "hgt_1", "hgt_3",
"wgt_0", "wgt_1", "wgt_2", "junk")

# note that it does not return hgt2
make_xname(yname = "hgt", xnames = xnames, user_model = 2,
current_age = 2)
}
