% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\title{Obtain an empty gTree object that can act as a temporary placeholder}
\usage{
placeholder(name = "")
}
\arguments{
\item{name}{A string indicating the name of the \code{gTree} object}
}
\value{
An object of class \code{gTree} with name \code{name}
}
\description{
Obtain an empty gTree object that can act as a temporary placeholder
}
