% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_target_height.R
\name{plot_target_height}
\alias{plot_target_height}
\title{Plot target height}
\usage{
plot_target_height(child, yname, parsed, options)
}
\arguments{
\item{child}{A data frame with columns}

\item{yname}{Variable code for vertical axis}

\item{parsed}{A list with parsed arguments}

\item{options}{A list with options (\code{show} and \code{alpha})}
}
\value{
A list of grobs
}
\description{
Calculate and plot target height and target height range for a child on
the growth charts with designs C and D.
}
