% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_flatchart.R
\name{process_flatchart}
\alias{process_flatchart}
\title{Plots the flat chart using ggplot or plotly}
\usage{
process_flatchart(
  target,
  chartcodes,
  plot_library = c("plotly", "ggplot2"),
  verbose = FALSE,
  draw_grob = FALSE,
  ...
)
}
\arguments{
\item{target}{A list with elements \code{"psn"} and \code{"xyz"}, e.g. as produced by
\code{bdsreader::read_bds()}}

\item{chartcodes}{A vector of chartcodes from which traces should be extracted}

\item{plot_library}{Either \code{plotly} (default) or \code{ggplot2}}

\item{verbose}{Logical. Set to \code{TRUE} to track down why traces may be missing}

\item{draw_grob}{Logical. Should chart be plotted on current device?
Default is \code{FALSE}. For internal use only.}

\item{\dots}{Arguments passed down to \code{centile::y2z()}}
}
\value{
Object of class \code{plotly} or \code{gg}
}
\description{
Plots the flat chart using ggplot or plotly
}
\examples{
\dontrun{
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json", package = "jamesdemodata")
target <- bdsreader::read_bds(fn)
g <- process_flatchart(target, chartcodes = "NMAA", plot_library = "plotly")
g
}
}
