% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_curves.R
\name{set_curves}
\alias{set_curves}
\title{Set data points in a graphic object}
\usage{
set_curves(
  g,
  target,
  ynames,
  curve_interpolation = TRUE,
  nmatch = 0L,
  matches = NULL,
  dnr = NULL,
  period = numeric(0),
  show_realized = FALSE,
  show_future = FALSE,
  clip = TRUE
)
}
\arguments{
\item{g}{A \code{gTree} object, typically loaded by
\code{chartbox::load_chart()}}

\item{target}{A list with elements \code{"psn"} and \code{"xyz"}, e.g. as produced by
\code{bdsreader::read_bds()}}

\item{ynames}{A vector with the names of the response variables
for which matches are sought, e.g. \code{ynames = c("hdc", "hgt")}.}

\item{curve_interpolation}{A logical indicating whether curve
interpolation shoud be applied.}

\item{nmatch}{Integer. Number of matches needed. When
\code{nmatch == 0L} no matches are sought.}

\item{matches}{List of vector containing the id's of the matches in
the donor data. List elements should be named after
\code{ynames}. The default value (\code{NULL}) indicates that no
matches are plotted.}

\item{dnr}{A string with the name of the donor data
(currently available are \code{smocc}, \code{terneuzen},
\code{lollypop} and \code{pops})}

\item{period}{A vector of length 2 with left and right ages
(decimal age). If \code{length(period) == 0L}, then no curve
matching is done}

\item{show_realized}{A logical indicating whether the realized
growth of the target child should be drawn}

\item{show_future}{A logical indicating whether the predicted
growth of the target child should be drawn}

\item{clip}{A logical indicating whether clipping is needed}
}
\value{
The grid object \code{g} with data points added.
}
\description{
Sets the data points in a graphic object. The function will identify
and set all \code{ynames} in the graph.
}
