% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits.R
\name{visit.label}
\alias{visit.label}
\title{Find the visit label given the visit number and donor data name}
\usage{
visit.label(number, dnr = "smocc")
}
\arguments{
\item{number}{Integer, visit numbers}

\item{dnr}{Name of donor dataset}
}
\value{
A vector of length \code{length(number)}
}
\description{
Find the visit label given the visit number and donor data name
}
\examples{
visit.label(1:3)
visit.label(1:3, "pops")
}
