% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits.R
\name{visit.number}
\alias{visit.number}
\title{Find the visit number given age and donor data name}
\usage{
visit.number(age, dnr = "smocc")
}
\arguments{
\item{age}{Numeric vector with decimal age}

\item{dnr}{Name of donor dataset}
}
\value{
A vector of length \code{length(age)}
}
\description{
Find the visit number given age and donor data name
}
\examples{
visit.number(age = c(0.5, 1))
visit.number(age = c(0.5, 1), "pops")
}
