% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits.R
\name{visit_number}
\alias{visit_number}
\alias{visit_age}
\alias{visit_label}
\title{Find the visit number given age and donor data name}
\usage{
visit_number(age, dnr = "smocc")

visit_age(number, dnr = "smocc")

visit_label(number, dnr = "smocc")
}
\arguments{
\item{age}{Numeric vector with decimal age}

\item{dnr}{Name of donor dataset}

\item{number}{Integer, visit numbers}
}
\value{
\code{visit_number} returns a vector of length \code{length(age)}

\code{visit_age} returns a vector of length \code{length(number)}

\code{visit_label} returns a vector of length \code{length(number)}
}
\description{
Find the visit number given age and donor data name

Find the visit age given the visit number and donor data name

Find the visit label given the visit number and donor data name
}
\examples{
chartplotter:::visit_number(age = c(0.5, 1))
chartplotter:::visit_number(age = c(0.5, 1), "pops")
chartplotter:::visit_age(1:3)
chartplotter:::visit_age(1:3, "pops")
chartplotter:::visit_label(1:3)
chartplotter:::visit_label(1:3, "pops")
}
