% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_interpolation.R
\name{curve_interpolation}
\alias{curve_interpolation}
\title{Interpolates the growth curves along the references}
\usage{
curve_interpolation(data, xname = "x", yname = "y", zname = NULL,
  xout, reference)
}
\arguments{
\item{data}{A data frame in long format, with columns for person
number (named \code{id}), the x and y variables.}

\item{xname}{Name of the variable on the horizontal axis.}

\item{yname}{Name of the variable on the vertical axis.}

\item{zname}{Name of the variable with Z-scores}

\item{xout}{A vector of values for the horizontal axis in the scale
of the variable \code{xname}. Only \code{xout} points within the
data range of \code{xname} are interpolated.}

\item{reference}{The reference needed to calculate the Z-score
scores and the back-transformation to measured scores.
An object of class \code{clopus::reference}.}
}
\value{
A \code{tibble} with five columns: \code{id}, xname, yname,
zname and \code{obs}. The \code{obs} variables signals whether
the point is observed or not.
}
\description{
We would normally connect the points of a growth curve by
straight lines. This is OK when ages are relatively close
together. However, for ages that are more apart, the straight
line gives an ugly representation that - moreover - underestimates
real growth during that interval.
}
\details{
The \code{curve_interpolation()} function provides a solution
for this. It defines a number of additional points (specified
through the \code{xout} argument), and calculates the linear
interpolation between these points in the Z-scores metric. After
that, the Z-scores are transformed back to the original scale,
so the points to plot then follow the curvy reference lines.
}
\examples{
data <- data.frame(
 id = c(1, 1, 1, 2, 3, 3, 3),
 age = c(0, 0.2, 1.2, 0.5, 0.1, 1, 1.3),
 hgt = c(52, 60, 78, 69, 62, 78, 82))
ref <- clopus::nl2009[["nl2009.mhgtNL"]]
xout <- seq(0, 3, 0.2)
int <- curve_interpolation(data, xname = "age", yname = "hgt", xout = xout, reference = ref)
int
}
\seealso{
\code{\link[clopus]{reference-class}}
}
\author{
Stef van Buuren, 2019
}
