% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_curves.R
\name{set_curves}
\alias{set_curves}
\title{Set data points in a graphic object}
\usage{
set_curves(g, individual, curve_interpolation = TRUE, matches = NULL,
  dnr = NULL, period = numeric(0), show_realized = FALSE,
  show_future = FALSE, clip = TRUE)
}
\arguments{
\item{g}{A \code{gTree} object, typically loaded by
\code{chartbox::load_chart()}}

\item{individual}{An S4 object of class \code{individual} containing data of the individual}

\item{curve_interpolation}{Flag indicating whether curves should bend to
reference curves}

\item{matches}{List of vector containing the id's of the matches in the donor data.
List elements should be named after \code{ynames}.
The default value (\code{NULL}) indicates that no matches are plotted.}

\item{dnr}{Name of the donor data (currently available are
\code{smocc}, \code{terneuzen}, \code{lollypop.preterm} or \code{lollypop.term})}

\item{period}{A vector of length 2 with left and right ages. If
\code{length(period) == 0L}, then no matches are plotted}

\item{show_realized}{A logical indicating whether the realized growth of the target
child should be drawn}

\item{show_future}{A logical indicating whether the predicted growth of the target
child should be drawn}

\item{clip}{A logical indicating whether clipping is needed}
}
\value{
The grid object \code{g} with data points added.
}
\description{
Sets the data points in a graphic object. The function will identify
and set all \code{ynames} in the graph.
}
